% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irps_class.R
\name{new_irps}
\alias{new_irps}
\title{Create an irps object}
\usage{
new_irps(
  ssgsea_scores,
  irps_scores,
  gene_sets,
  favorable_sets,
  unfavorable_sets,
  platform,
  n_features_used,
  params = list()
)
}
\arguments{
\item{ssgsea_scores}{Matrix of ssGSEA scores (gene sets × samples).}

\item{irps_scores}{Data frame of per-sample IRPS scores.}

\item{gene_sets}{Named list of gene sets used.}

\item{favorable_sets}{Names of favorable gene sets.}

\item{unfavorable_sets}{Names of unfavorable gene sets.}

\item{platform}{Character string: "proteomics" or "transcriptomics".}

\item{n_features_used}{Integer, number of features detected.}

\item{params}{List of additional parameters.}
}
\value{
An S3 object of class \code{irps}.
}
\description{
Create an irps object
}
\keyword{internal}
